// src/models/Requisiciones/plantaIdeal.ts
export type plantaIdeal = {
  Id?: string,
  Title: string;
  Tienda: string;
  "_x0031_": number;
  "_x0032_": number;
  "_x0033_": number;
  "_x0034_": number;
  "_x0035_": number;
  "_x0036_": number;
  "_x0037_": number;
  "_x0038_": number;
  "_x0039_": number;
  "_x0031_0": number;
  "_x0031_1": number;
  "_x0031_2": number;
};